+++ packages/front-end/src/components/optionsTrading/Modals/CloseOptionModal/components/Disclaimer.tsx
