+++ packages/front-end/src/components/optionsTrading/Modals/CloseOptionModal/components/Header.tsx
