+++ packages/front-end/src/components/optionsTrading/Modals/CloseOptionModal/components/Pricing.tsx
