+++ packages/front-end/src/components/optionsTrading/Modals/CloseOptionModal/utils/sell.ts
